The type regoff_t is defined as a signed arithmetic type that can hold the largest value that can be stored in either a type off_t or type ssize_t.
off_t does not have a specified range, therfore it may be larger than int.

Therfore regoff_t, if it is to be able to hold the largest value that can be stored in off_t, must also have an unspecified range.

Therfore regoff_t should indeed have ( signed ), not ( int ).

This is discussed in ticket #65.

svn path=/tendra/trunk/; revision=817

+++ tendra/src/lib/apis/posix2/regex.h
@@ -36,7 +36,7 @@
     size_t re_nsub ;
 } ;
 
-+TYPE ( int ) regoff_t ;
++TYPE ( signed ) regoff_t ;
 
 +FIELD ( struct ) regmatch_t {
     regoff_t rm_so ;
